+++ HISTORY.md
@@ -8,6 +8,7 @@
 * In locale files, the `general` and `national` parts are now merged in a new `sanctoral` part. This will avoid potential duplicated or missing keys between these 2 parts, particularly when a celebration is move from the general to the national calendar. Also, keys in celebrations and the new sanctoral have been sorted alphabetically. ([#97](https://github.com/romcal/romcal/pull/97))
 * Italian translation has been added in romcal ([#94](https://github.com/romcal/romcal/issues/94)).
 * In Christian calendars, Sunday is the first day of the week ([#103](https://github.com/romcal/romcal/issues/103))
+* Drop Shrove Monday and Tuesday from the general calendar ([#90](https://github.com/romcal/romcal/issues/90)).
 * Potential breaking changes:
     - The `WEEKDAY` type is renamed to `FERIA` ([#88](https://github.com/romcal/romcal/issues/88))
 * Fixes: